refactor(traces): extract shared updateTraceMetadata service (de-dup tRPC + REST span-synthesis, move logic out of transport)
